fix!: close the Android gates that did not need a device
Four changes that together finish everything the Android components can
complete inside this repository.

Two change emitted meaning and are breaking:

`geolocation.locationAgeMs` is a 64-bit age. The collector narrowed the
`Long` difference to an `Int`, so a fix older than ~24.86 days wrapped: a
30-day-old fix was emitted as -1702967296. The serialized JSON type is
unchanged, but the Kotlin signature is source-breaking for standalone
hosts, and any rule treating a negative age as suspect is now dead code.
The React Native converter already routed `Long` through `putDouble`, and
iOS never shared the defect because `NSInteger` is 64-bit on every
supported target; its cast now says `int64_t` so the width is stated.

The active probe's frida handshake is reassembled across reads, and
connect and read have separate budgets. One 6-byte read using the connect
timeout meant a `REJECT` split across segments, trickled out, or arriving
late read as "no REJECT". `fridaHandshakeReject` can now be true where it
was false, so values from before and after must not be compared. The
worst case per scan is a documented 1500 ms. The collapsed-false defect is
untouched and still recorded.

Two do not change behavior:

`DeviceRiskActiveProbes` declares `FRIDA_SCAN_ENABLED_BY_DEFAULT = true`
instead of leaving a native host to infer it. The AGENTS.md exemption is
written out - the probe is not new, it shipped enabled before extraction
and ADR-0003 preserved that default - along with the fact that
physical-device QA has not happened. A test pins the constant to the
catalog source, with that source declared as a task input so the pin
cannot go stale behind an up-to-date check.

The transaction session gains a 15-test instrumented suite covering no
activity, background-thread lifecycle calls on the real Looper, repeated
attach, activity switch and destroy, wrappers nested before and after
attach, stale generations, the granted-permission path, the API 29/34/35
gates and a three-reader snapshot race. Executed on an API 35 arm64
emulator; CI compiles it but has no device. It also pins a platform
finding: attaching to a destroyed Activity reports screenshot coverage it
cannot have, because API 35 does not reject the registration.

Both components now publish a Dokka-rendered javadoc jar. Signing,
Sonatype and a real version stay open - they need credentials that do not
live here - as does physical-device QA.

+++ ios/GeolocationInfoProvider.m
@@ -63,7 +63,10 @@
       }
       NSTimeInterval age = -[location.timestamp timeIntervalSinceNow];
       if (age >= 0) {
-        result[@"locationAgeMs"] = @((NSInteger)(age * 1000));
+        // 64-bit on purpose, matching the Kotlin Long: a 32-bit millisecond age wraps past
+        // ~24.86 days, which a long-cached fix or a backwards device clock reaches routinely.
+        // NSInteger is already 64-bit on every supported (arm64) iOS target; int64_t says so.
+        result[@"locationAgeMs"] = @((int64_t)(age * 1000));
       }
     }
   }
